Allow aliases for directories and resolve aliased paths normally
This changes the `alias` option to apply to the directories of imported
files (such as alias the `utils` directory in `@import "utils/other";`).
It also allows aliased paths to be processed with the full resolve
algorithm, for example importing an aliased directory will search for
the `index.css` file in the directory.

Also expanded the options documentation